plucker: fixed seg.fault html: fixed compilation problem
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@1225 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/common/formats/html/HtmlReader.cpp
@@ -53,8 +53,8 @@ void HtmlReader::setTag(HtmlTag &tag, const std::string &name) {
     tag.Name = name.substr(1);
   }
 
-  const unsigned int len = std::min(tag.Name.length(), (unsigned int)10);
-  for (unsigned int i = 0; i < len; i++) {
+  const size_t len = std::min(tag.Name.length(), (size_t)10);
+  for (size_t i = 0; i < len; i++) {
     tag.Name[i] = toupper(tag.Name[i]);
   }
 }
